requirements.txt
python-dotenv
openai
marker-pdf
litellm
